Added player protocol - runner needs fixing
